+++ WoltZajednicki/src/model/StavkaNarudzbine.java
@@ -62,7 +62,7 @@ public class StavkaNarudzbine extends GenericDomainObject {
                     rs.getString("a.Ime"), rs.getString("a.Prezime"),
                     rs.getString("Username"), rs.getString("Password"));
 
-            Mesto m = new Mesto(rs.getLong("mestoID"),
+            Mesto m = new Mesto(rs.getInt("mestoID"),
                     rs.getString("m.naziv"));
 
             Klijent k = new Klijent(rs.getLong("KlijentID"),
